feat: implement update stats by open modal
