+++ src/js/characters.js
@@ -1,14 +1,13 @@
 import { state } from './state';
 import initOverlayClose from './modal-overlay';
 import { resetFightByCharacterChange } from './fight';
+import { updateStatsUI } from './ui';
 
 const initChars = () => {
   const body = document.querySelector('body');
   const logo = document.querySelector('.logo');
   const buttonClose = document.querySelector('.characters__button-close');
   const nicknames = document.querySelectorAll('.characters__nickname');
-  const wins = document.querySelectorAll('.characters__result-w');
-  const loses = document.querySelectorAll('.characters__result-l');
   const selectedChars = document.querySelectorAll('.characters__selected');
   const settingsLink = document.querySelector('a[href="#settings"]');
   const attackButtons = document.querySelectorAll(
@@ -18,19 +17,15 @@ const initChars = () => {
     '.defence-panel .fight__button',
   );
 
+  const updateStatsOnModalOpen = () => {
+    if (location.hash === '#characters') {
+      updateStatsUI();
+    }
+  };
+
   if (state.nickname) {
     nicknames.forEach((nickname) => {
       nickname.innerHTML = state.nickname;
-    });
-
-    state.characters.forEach((char, index) => {
-      if (wins[index]) {
-        wins[index].innerHTML = char.stats.win;
-      }
-
-      if (loses[index]) {
-        loses[index].innerHTML = char.stats.lose;
-      }
     });
 
     const defaultChar = state.avatarId ?? 0;
@@ -92,6 +87,8 @@ const initChars = () => {
   }
 
   initOverlayClose(closeModal);
+  updateStatsOnModalOpen();
+  window.addEventListener('hashchange', updateStatsOnModalOpen);
 };
 
 export default initChars;
